[Architecture] Switch over to a common impl with separate drivers
The functionality is the same, but now all personalities have a common wrapper around a driver instead of having separate implementations directly in their hand.

This is a bit more convoluted than seems necessary, but the layers of
indirection allow us to properly hide the implementation methods because
they're only available via a class that is internal.

TODO: roll this out to the other implementations